просто привел код в порядок не реализовал удаление чекбоксами, просто добавил алерт при нажатии  кнопки удалить

+++ testovoe/resources/js/components/projects/ProjectsIndex.vue
@@ -3,9 +3,9 @@
     <table class="table">
       <thead>
         <tr>
-          <th scope="col">Name</th>
-          <th scope="col">Description</th>
-          <th scope="col">Actions</th>
+          <th scope="col">Имя</th>
+          <th scope="col">Описание</th>
+          <th scope="col">Действия</th>
         </tr>
       </thead>
       <tbody>
@@ -15,10 +15,10 @@
           <td>
             <div class="row gap-3">
               <router-link :to="{ name: 'ProjectsCreate', params: { id: project.id } }" class="p-2 col border btn btn-success">
-                Edit
+                Изменить
               </router-link>
               <button @click="deleteProject(project.id)" type="button"
-                class="p-2 col border btn btn-danger">Delete</button>
+                class="p-2 col border btn btn-danger">Удалить</button>
             </div>
           </td>
         </tr>
@@ -47,8 +47,10 @@ export default {
   methods: {
     async deleteProject(id) {
       try {
+        if (confirm("Удалить выбранный проект?")) {
         await axios.delete(`/api/projectgal/${id}`);
         this.projects = this.projects.filter(project => project.id !== id);
+        }
       } catch (error) {
         console.error(error.response.data);
       }
